fix: Fix recently answered quiz item being asked again
`createNewQuiz()` of useTimesTableQuiz hook will now only be called via
`startQuiz()` or by a useEffect whenever `quizResults` changes in length
which always gets incremented due to every answer submitted by the user
to populate `quizResults`.

The previous problem caused by react states nature of being
asynchronous, that is why the check for recent answered item is not
consistent resulting to not preventing reduplicated quiz items. Using a
`useEffect` will do the job as it can see the latest changes of
`quizResults` as it gets called only when there is a new item added to
the array.

The logic for checking equality between the new generated quiz and
recent answered quiz has been fixed, because the previous logic just
checks for reference of the objects not the actual value of their
multiplicand and multiplier properties.