fix: Fix toggle theme from not working when clicked at the first time
The themeSetting local storage item is not being updated properly
because of the usage of setThemeSetting depends on it's previous
property which is actually null causing it to fail the ternary operator
which sets it with a value of "light". This makes no visual impact if
the app is currently in light mode as the theme is being applied with
the same theme.

To fix this, the setThemeSetting should only be called whenever the
theme state (the in app state) changes.